+++ tests/misc/tests/lamports/lamports.ts
@@ -8,16 +8,13 @@ describe("lamports", () => {
 
   const program = anchor.workspace.Lamports as anchor.Program<Lamports>;
 
-  it("Can use the Lamports trait", async () => {
-    const signer = program.provider.publicKey!;
-    const [pda] = anchor.web3.PublicKey.findProgramAddressSync(
-      [Buffer.from("lamports")],
-      program.programId
-    );
-
+  it("Can transfer from/to PDA", async () => {
     await program.methods
-      .testLamportsTrait(new anchor.BN(anchor.web3.LAMPORTS_PER_SOL))
-      .accounts({ signer, pda })
+      .transfer(new anchor.BN(anchor.web3.LAMPORTS_PER_SOL))
       .rpc();
+  });
+
+  it("Returns an error on overflow", async () => {
+    await program.methods.overflow().rpc();
   });
 });